requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonfield==3.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.2
openapi-codec==1.3.2
packaging==20.3
pandas==1.0.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
gunicorn
psycopg2-binary
whitenoise==3.3.1
